requirements-prod.txt
-r requirements.txt
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==1.4.41
typing-extensions==4.7.1 